backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3
scikit-learn==1.4.0
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
